refactor: improve WalletAssetsModal structure and remove evm fundingAsset on disconnect wallet
